Fixing Route

+++ app/src/main/java/com/herbify/herbifyapp/model/UserPreferences.kt
@@ -37,7 +37,7 @@ class UserPreferences private constructor(context: Context){
 
     fun getUser(): UserModel{
         return UserModel(
-            id = preferences.getInt(ID_KEY, 0),
+            id = preferences.getLong(ID_KEY, 0),
             name = preferences.getString(NAME_KEY, null),
             token = preferences.getString(TOKEN_KEY, null),
             email = preferences.getString(EMAIL_KEY, null),
@@ -55,9 +55,9 @@ class UserPreferences private constructor(context: Context){
         preferences.edit().putBoolean(VERIFIED_KEY, true).apply()
     }
 
-    fun login(name: String, email: String, id: Int, token: String, isVerified: Boolean){
+    fun login(name: String, email: String, id: Long, token: String, isVerified: Boolean){
         preferences.edit()
-            .putInt(ID_KEY, id)
+            .putLong(ID_KEY, id)
             .putString(NAME_KEY, name)
             .putString(EMAIL_KEY,email)
             .putString(TOKEN_KEY, token)
